[UPDATE] endpoint architecture
