+++ app/back/server/src/lib/components/endpoints/__init__.py
@@ -1,0 +1,10 @@
+"""
+"""
+
+from .authentication import Authentication
+from .bonus import Bonus
+
+__all__ = [
+    "Authentication",
+    "Bonus"
+]
